Changed the filter panel to a modal
